+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 * [MIC-8723] - Design review ui fixes - Pay by new card
 * [MIC-8719] - Design review ui fixes - Cards flow
 * [MIC-8786] Fix keyboard handling in recurrent payment module
+* [EACQAPW-4284] Testing SBPBanks module and services
 
 ## [2.19.0] - 2023-04-13Z